feat(02-02): split server.py into tools/ package with domain modules
- Create tools/ package: session, tracks, clips, transport, devices, browser + mixer/scenes placeholders
- Rewrite server.py as slim 48-line orchestrator (FastMCP instance + lifespan + imports)
- Add shutdown_connection() to connection.py for clean lifecycle management
- Update __init__.py to re-export from connection module
- Update pyproject.toml to include MCP_Server.tools sub-package
- Update test fixtures to point at correct post-refactor source files
- All 17 MCP tools registered and verified via list_tools()
