+++ MCP_Server/server.py
@@ -1,183 +1,20 @@
-# ableton_mcp_server.py
-from mcp.server.fastmcp import FastMCP, Context
-import socket
-import struct
-import json
+"""AbletonMCP server -- MCP bridge to Ableton Live."""
 import logging
-import threading
-import time
-from dataclasses import dataclass
 from contextlib import asynccontextmanager
-from typing import AsyncIterator, Dict, Any, List, Union
+from typing import AsyncIterator, Dict, Any
+
+from mcp.server.fastmcp import FastMCP
+from MCP_Server.connection import get_ableton_connection, shutdown_connection
 
 # Configure logging
 logging.basicConfig(level=logging.INFO,
                     format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
 logger = logging.getLogger("AbletonMCPServer")
 
-# Timeout constants for different operation types
-TIMEOUT_READ = 10.0
-TIMEOUT_WRITE = 15.0
-TIMEOUT_BROWSER = 30.0
-TIMEOUT_PING = 5.0
-
-# Commands that require longer timeouts (browser/load operations)
-_BROWSER_COMMANDS = frozenset([
-    "get_browser_tree", "get_browser_items_at_path", "get_browser_item",
-    "get_browser_categories", "get_browser_items",
-    "load_browser_item", "load_instrument_or_effect",
-])
-
-# Commands that modify state (write operations)
-_WRITE_COMMANDS = frozenset([
-    "create_midi_track", "create_audio_track", "set_track_name",
-    "create_clip", "add_notes_to_clip", "set_clip_name",
-    "set_tempo", "fire_clip", "stop_clip", "set_device_parameter",
-    "start_playback", "stop_playback",
-])
-
-
-# --- AI-friendly error formatting ---
-
-def format_error(message: str, detail: str = "", suggestion: str = "") -> str:
-    """Format error for AI consumption. Clean message first, technical detail below."""
-    parts = [f"Error: {message}"]
-    if suggestion:
-        parts.append(f"Suggestion: {suggestion}")
-    if detail:
-        parts.append(f"Debug: {detail}")
-    return "\n".join(parts)
-
-
-# --- Length-prefix framing protocol ---
-
-def _recv_exact(sock: socket.socket, n: int):
-    """Read exactly n bytes from socket."""
-    buf = bytearray()
-    while len(buf) < n:
-        chunk = sock.recv(n - len(buf))
-        if not chunk:
-            return None
-        buf.extend(chunk)
-    return bytes(buf)
-
-
-def send_message(sock: socket.socket, data: dict) -> None:
-    """Send a length-prefixed JSON message."""
-    payload = json.dumps(data, ensure_ascii=False).encode("utf-8")
-    header = struct.pack(">I", len(payload))
-    sock.sendall(header + payload)
-
-
-def recv_message(sock: socket.socket, timeout: float = 15.0) -> dict:
-    """Receive a length-prefixed JSON message."""
-    sock.settimeout(timeout)
-    header = _recv_exact(sock, 4)
-    if not header:
-        raise ConnectionError("Connection closed while reading header")
-    length = struct.unpack(">I", header)[0]
-    if length > 10 * 1024 * 1024:  # 10MB safety limit
-        raise ValueError(f"Message too large: {length} bytes")
-    payload = _recv_exact(sock, length)
-    if not payload:
-        raise ConnectionError("Connection closed while reading payload")
-    return json.loads(payload.decode("utf-8"))
-
-
-def _timeout_for(command_type: str) -> float:
-    """Return the appropriate timeout for a command type."""
-    if command_type == "ping":
-        return TIMEOUT_PING
-    if command_type in _BROWSER_COMMANDS:
-        return TIMEOUT_BROWSER
-    if command_type in _WRITE_COMMANDS:
-        return TIMEOUT_WRITE
-    return TIMEOUT_READ
-
-
-@dataclass
-class AbletonConnection:
-    host: str
-    port: int
-    sock: socket.socket = None
-
-    def connect(self) -> bool:
-        """Connect to the Ableton Remote Script socket server"""
-        if self.sock:
-            return True
-
-        try:
-            self.sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-            self.sock.connect((self.host, self.port))
-            logger.info(f"Connected to Ableton at {self.host}:{self.port}")
-            return True
-        except Exception as e:
-            logger.error(f"Failed to connect to Ableton: {str(e)}")
-            self.sock = None
-            return False
-
-    def disconnect(self):
-        """Disconnect from the Ableton Remote Script"""
-        if self.sock:
-            try:
-                self.sock.close()
-            except Exception as e:
-                logger.error(f"Error disconnecting from Ableton: {str(e)}")
-            finally:
-                self.sock = None
-
-    def send_command(self, command_type: str, params: Dict[str, Any] = None) -> Dict[str, Any]:
-        """Send a command to Ableton and return the response.
-
-        Uses length-prefix framing for reliable message boundaries.
-        No artificial delays -- the framing protocol handles completeness.
-        """
-        if not self.sock and not self.connect():
-            raise ConnectionError("Not connected to Ableton")
-
-        command = {
-            "type": command_type,
-            "params": params or {}
-        }
-
-        timeout = _timeout_for(command_type)
-
-        try:
-            logger.info(f"Sending command: {command_type} with params: {params}")
-
-            # Send the command using length-prefix framing
-            send_message(self.sock, command)
-            logger.info("Command sent, waiting for response...")
-
-            # Receive the response using length-prefix framing
-            response = recv_message(self.sock, timeout=timeout)
-            logger.info(f"Response received, status: {response.get('status', 'unknown')}")
-
-            if response.get("status") == "error":
-                logger.error(f"Ableton error: {response.get('message')}")
-                raise Exception(response.get("message", "Unknown error from Ableton"))
-
-            return response.get("result", {})
-        except socket.timeout:
-            logger.error("Socket timeout while waiting for response from Ableton")
-            self.sock = None
-            raise Exception("Timeout waiting for Ableton response")
-        except (ConnectionError, BrokenPipeError, ConnectionResetError) as e:
-            logger.error(f"Socket connection error: {str(e)}")
-            self.sock = None
-            raise Exception(f"Connection to Ableton lost: {str(e)}")
-        except json.JSONDecodeError as e:
-            logger.error(f"Invalid JSON response from Ableton: {str(e)}")
-            self.sock = None
-            raise Exception(f"Invalid response from Ableton: {str(e)}")
-        except Exception as e:
-            logger.error(f"Error communicating with Ableton: {str(e)}")
-            self.sock = None
-            raise Exception(f"Communication error with Ableton: {str(e)}")
 
 @asynccontextmanager
 async def server_lifespan(server: FastMCP) -> AsyncIterator[Dict[str, Any]]:
-    """Manage server startup and shutdown lifecycle"""
+    """Manage Ableton connection lifecycle."""
     try:
         logger.info("AbletonMCP server starting up")
 
@@ -190,608 +27,22 @@ async def server_lifespan(server: FastMCP) -> AsyncIterator[Dict[str, Any]]:
 
         yield {}
     finally:
-        global _ableton_connection
-        with _connection_lock:
-            if _ableton_connection:
-                logger.info("Disconnecting from Ableton on shutdown")
-                _ableton_connection.disconnect()
-                _ableton_connection = None
+        logger.info("Disconnecting from Ableton on shutdown")
+        shutdown_connection()
         logger.info("AbletonMCP server shut down")
 
+
 # Create the MCP server with lifespan support
-mcp = FastMCP(
-    "AbletonMCP",
-    lifespan=server_lifespan
-)
+mcp = FastMCP("AbletonMCP", lifespan=server_lifespan)
 
-# Global connection for resources -- protected by _connection_lock
-_ableton_connection = None
-_connection_lock = threading.Lock()
-
-def get_ableton_connection():
-    """Get or create a persistent Ableton connection.
-
-    Thread-safe: all access to _ableton_connection is serialized
-    by _connection_lock. Uses a real ping command for liveness testing
-    instead of sending empty bytes.
-    """
-    global _ableton_connection
-
-    with _connection_lock:
-        if _ableton_connection is not None:
-            try:
-                # Test the connection with a real ping command
-                _ableton_connection.send_command("ping")
-                return _ableton_connection
-            except Exception as e:
-                logger.warning(f"Existing connection is no longer valid: {str(e)}")
-                try:
-                    _ableton_connection.disconnect()
-                except Exception as cleanup_err:
-                    logger.warning(f"Error during connection cleanup: {cleanup_err}")
-                _ableton_connection = None
-
-        # Connection doesn't exist or is invalid, create a new one
-        # Try to connect up to 3 times with a short delay between attempts
-        max_attempts = 3
-        for attempt in range(1, max_attempts + 1):
-            try:
-                logger.info(f"Connecting to Ableton (attempt {attempt}/{max_attempts})...")
-                _ableton_connection = AbletonConnection(host="localhost", port=9877)
-                if _ableton_connection.connect():
-                    logger.info("Created new persistent connection to Ableton")
-
-                    # Validate connection with a simple command
-                    try:
-                        _ableton_connection.send_command("get_session_info")
-                        logger.info("Connection validated successfully")
-                        return _ableton_connection
-                    except Exception as e:
-                        logger.error(f"Connection validation failed: {str(e)}")
-                        _ableton_connection.disconnect()
-                        _ableton_connection = None
-                        # Continue to next attempt
-                else:
-                    _ableton_connection = None
-            except Exception as e:
-                logger.error(f"Connection attempt {attempt} failed: {str(e)}")
-                if _ableton_connection:
-                    _ableton_connection.disconnect()
-                    _ableton_connection = None
-
-            # Wait before trying again, but only if we have more attempts left
-            if attempt < max_attempts:
-                time.sleep(1.0)
-
-        # If we get here, all connection attempts failed
-        if _ableton_connection is None:
-            logger.error("Failed to connect to Ableton after multiple attempts")
-            raise Exception("Could not connect to Ableton. Make sure the Remote Script is running.")
-
-    return _ableton_connection
+# Import tool modules AFTER mcp is created (triggers @mcp.tool() registration)
+import MCP_Server.tools  # noqa: E402, F401
 
 
-# Core Tool endpoints
-
-@mcp.tool()
-def get_connection_status(ctx: Context) -> str:
-    """Check connection health and get Ableton session summary.
-    Returns connection state, Ableton version, and basic session info.
-    Use this to verify the connection before starting work."""
-    try:
-        ableton = get_ableton_connection()
-        ping_result = ableton.send_command("ping")
-        session = ableton.send_command("get_session_info")
-        return json.dumps({
-            "connected": True,
-            "ableton_version": ping_result.get("ableton_version", "unknown"),
-            "tempo": session.get("tempo"),
-            "track_count": session.get("track_count"),
-            "return_track_count": session.get("return_track_count"),
-        }, indent=2)
-    except Exception as e:
-        return format_error(
-            "Cannot reach Ableton",
-            detail=str(e),
-            suggestion="Ensure Ableton is running and Remote Script is loaded in Preferences > Link/Tempo/MIDI"
-        )
-
-
-@mcp.tool()
-def get_session_info(ctx: Context) -> str:
-    """Get detailed information about the current Ableton session"""
-    try:
-        ableton = get_ableton_connection()
-        result = ableton.send_command("get_session_info")
-        return json.dumps(result, indent=2)
-    except Exception as e:
-        logger.error(f"Error getting session info from Ableton: {str(e)}")
-        return format_error(
-            "Failed to get session info",
-            detail=str(e),
-            suggestion="Verify connection with get_connection_status first"
-        )
-
-@mcp.tool()
-def get_track_info(ctx: Context, track_index: int) -> str:
-    """
-    Get detailed information about a specific track in Ableton.
-
-    Parameters:
-    - track_index: The index of the track to get information about
-    """
-    try:
-        ableton = get_ableton_connection()
-        result = ableton.send_command("get_track_info", {"track_index": track_index})
-        return json.dumps(result, indent=2)
-    except Exception as e:
-        logger.error(f"Error getting track info from Ableton: {str(e)}")
-        return format_error(
-            "Failed to get track info",
-            detail=str(e),
-            suggestion="Verify track_index is valid with get_session_info"
-        )
-
-@mcp.tool()
-def create_midi_track(ctx: Context, index: int = -1) -> str:
-    """
-    Create a new MIDI track in the Ableton session.
-
-    Parameters:
-    - index: The index to insert the track at (-1 = end of list)
-    """
-    try:
-        ableton = get_ableton_connection()
-        result = ableton.send_command("create_midi_track", {"index": index})
-        return f"Created new MIDI track: {result.get('name', 'unknown')}"
-    except Exception as e:
-        logger.error(f"Error creating MIDI track: {str(e)}")
-        return format_error(
-            "Failed to create MIDI track",
-            detail=str(e),
-            suggestion="Check track count with get_session_info"
-        )
-
-
-@mcp.tool()
-def set_track_name(ctx: Context, track_index: int, name: str) -> str:
-    """
-    Set the name of a track.
-
-    Parameters:
-    - track_index: The index of the track to rename
-    - name: The new name for the track
-    """
-    try:
-        ableton = get_ableton_connection()
-        result = ableton.send_command("set_track_name", {"track_index": track_index, "name": name})
-        return f"Renamed track to: {result.get('name', name)}"
-    except Exception as e:
-        logger.error(f"Error setting track name: {str(e)}")
-        return format_error(
-            "Failed to set track name",
-            detail=str(e),
-            suggestion="Verify track_index is valid with get_session_info"
-        )
-
-@mcp.tool()
-def create_clip(ctx: Context, track_index: int, clip_index: int, length: float = 4.0) -> str:
-    """
-    Create a new MIDI clip in the specified track and clip slot.
-
-    Parameters:
-    - track_index: The index of the track to create the clip in
-    - clip_index: The index of the clip slot to create the clip in
-    - length: The length of the clip in beats (default: 4.0)
-    """
-    try:
-        ableton = get_ableton_connection()
-        result = ableton.send_command("create_clip", {
-            "track_index": track_index,
-            "clip_index": clip_index,
-            "length": length
-        })
-        return f"Created new clip at track {track_index}, slot {clip_index} with length {length} beats"
-    except Exception as e:
-        logger.error(f"Error creating clip: {str(e)}")
-        return format_error(
-            "Failed to create clip",
-            detail=str(e),
-            suggestion="Verify track_index and clip_index with get_track_info"
-        )
-
-@mcp.tool()
-def add_notes_to_clip(
-    ctx: Context,
-    track_index: int,
-    clip_index: int,
-    notes: List[Dict[str, Union[int, float, bool]]]
-) -> str:
-    """
-    Add MIDI notes to a clip.
-
-    Parameters:
-    - track_index: The index of the track containing the clip
-    - clip_index: The index of the clip slot containing the clip
-    - notes: List of note dictionaries, each with pitch, start_time, duration, velocity, and mute
-    """
-    try:
-        ableton = get_ableton_connection()
-        result = ableton.send_command("add_notes_to_clip", {
-            "track_index": track_index,
-            "clip_index": clip_index,
-            "notes": notes
-        })
-        return f"Added {len(notes)} notes to clip at track {track_index}, slot {clip_index}"
-    except Exception as e:
-        logger.error(f"Error adding notes to clip: {str(e)}")
-        return format_error(
-            "Failed to add notes to clip",
-            detail=str(e),
-            suggestion="Verify clip exists at the specified track and slot with get_track_info"
-        )
-
-@mcp.tool()
-def set_clip_name(ctx: Context, track_index: int, clip_index: int, name: str) -> str:
-    """
-    Set the name of a clip.
-
-    Parameters:
-    - track_index: The index of the track containing the clip
-    - clip_index: The index of the clip slot containing the clip
-    - name: The new name for the clip
-    """
-    try:
-        ableton = get_ableton_connection()
-        result = ableton.send_command("set_clip_name", {
-            "track_index": track_index,
-            "clip_index": clip_index,
-            "name": name
-        })
-        return f"Renamed clip at track {track_index}, slot {clip_index} to '{name}'"
-    except Exception as e:
-        logger.error(f"Error setting clip name: {str(e)}")
-        return format_error(
-            "Failed to set clip name",
-            detail=str(e),
-            suggestion="Verify clip exists at the specified track and slot with get_track_info"
-        )
-
-@mcp.tool()
-def set_tempo(ctx: Context, tempo: float) -> str:
-    """
-    Set the tempo of the Ableton session.
-
-    Parameters:
-    - tempo: The new tempo in BPM
-    """
-    try:
-        ableton = get_ableton_connection()
-        result = ableton.send_command("set_tempo", {"tempo": tempo})
-        return f"Set tempo to {tempo} BPM"
-    except Exception as e:
-        logger.error(f"Error setting tempo: {str(e)}")
-        return format_error(
-            "Failed to set tempo",
-            detail=str(e),
-            suggestion="Tempo must be between 20 and 999 BPM"
-        )
-
-
-@mcp.tool()
-def load_instrument_or_effect(ctx: Context, track_index: int, uri: str) -> str:
-    """
-    Load an instrument or effect onto a track using its URI.
-
-    Parameters:
-    - track_index: The index of the track to load the instrument on
-    - uri: The URI of the instrument or effect to load (e.g., 'query:Synths#Instrument%20Rack:Bass:FileId_5116')
-    """
-    try:
-        ableton = get_ableton_connection()
-        result = ableton.send_command("load_browser_item", {
-            "track_index": track_index,
-            "item_uri": uri
-        })
-
-        # Check if the instrument was loaded successfully
-        if result.get("loaded", False):
-            item_name = result.get("item_name", "unknown")
-            device_chain = result.get("devices", [])
-            track_name = result.get("track_name", f"track {track_index}")
-            if device_chain:
-                return (
-                    f"Loaded '{item_name}' on {track_name} "
-                    f"(device chain: {', '.join(device_chain)})"
-                )
-            else:
-                return f"Loaded '{item_name}' on {track_name}"
-        else:
-            error_msg = result.get("error", "Unknown reason")
-            return format_error(
-                f"Failed to load instrument with URI '{uri}'",
-                detail=error_msg,
-                suggestion="Verify the URI using get_browser_items_at_path first"
-            )
-    except Exception as e:
-        logger.error(f"Error loading instrument by URI: {str(e)}")
-        return format_error(
-            "Failed to load instrument",
-            detail=str(e),
-            suggestion="Verify the URI using get_browser_items_at_path first"
-        )
-
-@mcp.tool()
-def fire_clip(ctx: Context, track_index: int, clip_index: int) -> str:
-    """
-    Start playing a clip.
-
-    Parameters:
-    - track_index: The index of the track containing the clip
-    - clip_index: The index of the clip slot containing the clip
-    """
-    try:
-        ableton = get_ableton_connection()
-        result = ableton.send_command("fire_clip", {
-            "track_index": track_index,
-            "clip_index": clip_index
-        })
-        return f"Started playing clip at track {track_index}, slot {clip_index}"
-    except Exception as e:
-        logger.error(f"Error firing clip: {str(e)}")
-        return format_error(
-            "Failed to fire clip",
-            detail=str(e),
-            suggestion="Verify clip exists at the specified track and slot with get_track_info"
-        )
-
-@mcp.tool()
-def stop_clip(ctx: Context, track_index: int, clip_index: int) -> str:
-    """
-    Stop playing a clip.
-
-    Parameters:
-    - track_index: The index of the track containing the clip
-    - clip_index: The index of the clip slot containing the clip
-    """
-    try:
-        ableton = get_ableton_connection()
-        result = ableton.send_command("stop_clip", {
-            "track_index": track_index,
-            "clip_index": clip_index
-        })
-        return f"Stopped clip at track {track_index}, slot {clip_index}"
-    except Exception as e:
-        logger.error(f"Error stopping clip: {str(e)}")
-        return format_error(
-            "Failed to stop clip",
-            detail=str(e),
-            suggestion="Verify clip exists at the specified track and slot with get_track_info"
-        )
-
-@mcp.tool()
-def start_playback(ctx: Context) -> str:
-    """Start playing the Ableton session."""
-    try:
-        ableton = get_ableton_connection()
-        result = ableton.send_command("start_playback")
-        return "Started playback"
-    except Exception as e:
-        logger.error(f"Error starting playback: {str(e)}")
-        return format_error(
-            "Failed to start playback",
-            detail=str(e),
-            suggestion="Verify connection with get_connection_status"
-        )
-
-@mcp.tool()
-def stop_playback(ctx: Context) -> str:
-    """Stop playing the Ableton session."""
-    try:
-        ableton = get_ableton_connection()
-        result = ableton.send_command("stop_playback")
-        return "Stopped playback"
-    except Exception as e:
-        logger.error(f"Error stopping playback: {str(e)}")
-        return format_error(
-            "Failed to stop playback",
-            detail=str(e),
-            suggestion="Verify connection with get_connection_status"
-        )
-
-@mcp.tool()
-def get_browser_tree(ctx: Context, category_type: str = "all") -> str:
-    """
-    Get a hierarchical tree of browser categories from Ableton.
-
-    Parameters:
-    - category_type: Type of categories to get ('all', 'instruments', 'sounds', 'drums', 'audio_effects', 'midi_effects')
-    """
-    try:
-        ableton = get_ableton_connection()
-        result = ableton.send_command("get_browser_tree", {
-            "category_type": category_type
-        })
-
-        # Check if we got any categories
-        if "available_categories" in result and len(result.get("categories", [])) == 0:
-            available_cats = result.get("available_categories", [])
-            return (f"No categories found for '{category_type}'. "
-                   f"Available browser categories: {', '.join(available_cats)}")
-
-        # Format the tree in a more readable way
-        total_folders = result.get("total_folders", 0)
-        formatted_output = f"Browser tree for '{category_type}' (showing {total_folders} folders):\n\n"
-
-        def format_tree(item, indent=0):
-            output = ""
-            if item:
-                prefix = "  " * indent
-                name = item.get("name", "Unknown")
-                path = item.get("path", "")
-                has_more = item.get("has_more", False)
-
-                # Add this item
-                output += f"{prefix}• {name}"
-                if path:
-                    output += f" (path: {path})"
-                if has_more:
-                    output += " [...]"
-                output += "\n"
-
-                # Add children
-                for child in item.get("children", []):
-                    output += format_tree(child, indent + 1)
-            return output
-
-        # Format each category
-        for category in result.get("categories", []):
-            formatted_output += format_tree(category)
-            formatted_output += "\n"
-
-        return formatted_output
-    except Exception as e:
-        error_msg = str(e)
-        if "Browser is not available" in error_msg:
-            logger.error(f"Browser is not available in Ableton: {error_msg}")
-            return format_error(
-                "Ableton browser is not available",
-                detail=error_msg,
-                suggestion="Make sure Ableton Live is fully loaded and try again"
-            )
-        elif "Could not access Live application" in error_msg:
-            logger.error(f"Could not access Live application: {error_msg}")
-            return format_error(
-                "Could not access Ableton Live application",
-                detail=error_msg,
-                suggestion="Make sure Ableton Live is running and Remote Script is loaded"
-            )
-        else:
-            logger.error(f"Error getting browser tree: {error_msg}")
-            return format_error(
-                "Failed to get browser tree",
-                detail=error_msg,
-                suggestion="Verify connection with get_connection_status"
-            )
-
-@mcp.tool()
-def get_browser_items_at_path(ctx: Context, path: str) -> str:
-    """
-    Get browser items at a specific path in Ableton's browser.
-
-    Parameters:
-    - path: Path in the format "category/folder/subfolder"
-            where category is one of the available browser categories in Ableton
-    """
-    try:
-        ableton = get_ableton_connection()
-        result = ableton.send_command("get_browser_items_at_path", {
-            "path": path
-        })
-
-        # Check if there was an error with available categories
-        if "error" in result and "available_categories" in result:
-            error = result.get("error", "")
-            available_cats = result.get("available_categories", [])
-            return (f"Error: {error}\n"
-                   f"Available browser categories: {', '.join(available_cats)}")
-
-        return json.dumps(result, indent=2)
-    except Exception as e:
-        error_msg = str(e)
-        if "Browser is not available" in error_msg:
-            logger.error(f"Browser is not available in Ableton: {error_msg}")
-            return format_error(
-                "Ableton browser is not available",
-                detail=error_msg,
-                suggestion="Make sure Ableton Live is fully loaded and try again"
-            )
-        elif "Could not access Live application" in error_msg:
-            logger.error(f"Could not access Live application: {error_msg}")
-            return format_error(
-                "Could not access Ableton Live application",
-                detail=error_msg,
-                suggestion="Make sure Ableton Live is running and Remote Script is loaded"
-            )
-        elif "Unknown or unavailable category" in error_msg:
-            logger.error(f"Invalid browser category: {error_msg}")
-            return format_error(
-                "Invalid browser category",
-                detail=error_msg,
-                suggestion="Check available categories using get_browser_tree"
-            )
-        elif "Path part" in error_msg and "not found" in error_msg:
-            logger.error(f"Path not found: {error_msg}")
-            return format_error(
-                "Browser path not found",
-                detail=error_msg,
-                suggestion="Check the path and try again"
-            )
-        else:
-            logger.error(f"Error getting browser items at path: {error_msg}")
-            return format_error(
-                "Failed to get browser items at path",
-                detail=error_msg,
-                suggestion="Verify connection with get_connection_status"
-            )
-
-@mcp.tool()
-def load_drum_kit(ctx: Context, track_index: int, rack_uri: str, kit_path: str) -> str:
-    """
-    Load a drum rack and then load a specific drum kit into it.
-
-    Parameters:
-    - track_index: The index of the track to load on
-    - rack_uri: The URI of the drum rack to load (e.g., 'Drums/Drum Rack')
-    - kit_path: Path to the drum kit inside the browser (e.g., 'drums/acoustic/kit1')
-    """
-    try:
-        ableton = get_ableton_connection()
-
-        # Step 1: Load the drum rack
-        result = ableton.send_command("load_browser_item", {
-            "track_index": track_index,
-            "item_uri": rack_uri
-        })
-
-        if not result.get("loaded", False):
-            return f"Failed to load drum rack with URI '{rack_uri}'"
-
-        # Step 2: Get the drum kit items at the specified path
-        kit_result = ableton.send_command("get_browser_items_at_path", {
-            "path": kit_path
-        })
-
-        if "error" in kit_result:
-            return f"Loaded drum rack but failed to find drum kit: {kit_result.get('error')}"
-
-        # Step 3: Find a loadable drum kit
-        kit_items = kit_result.get("items", [])
-        loadable_kits = [item for item in kit_items if item.get("is_loadable", False)]
-
-        if not loadable_kits:
-            return f"Loaded drum rack but no loadable drum kits found at '{kit_path}'"
-
-        # Step 4: Load the first loadable kit
-        kit_uri = loadable_kits[0].get("uri")
-        load_result = ableton.send_command("load_browser_item", {
-            "track_index": track_index,
-            "item_uri": kit_uri
-        })
-
-        return f"Loaded drum rack and kit '{loadable_kits[0].get('name')}' on track {track_index}"
-    except Exception as e:
-        logger.error(f"Error loading drum kit: {str(e)}")
-        return format_error(
-            "Failed to load drum kit",
-            detail=str(e),
-            suggestion="Verify the rack_uri and kit_path using get_browser_items_at_path first"
-        )
-
-# Main execution
 def main():
-    """Run the MCP server"""
+    """Run the MCP server."""
     mcp.run()
+
 
 if __name__ == "__main__":
     main()
