+++ MCP_Server/tools/__init__.py
@@ -1,0 +1,4 @@
+"""Import all tool modules to trigger @mcp.tool() registration."""
+from . import session, tracks, clips, transport, devices, browser  # noqa: F401
+# Placeholder modules (no tools yet):
+# from . import mixer, scenes
