+++ MCP_Server/tools/scenes.py
@@ -1,0 +1,1 @@
+"""Scene and transport tools -- populated in Phase 8 (SCNE-01..04)."""
